requirements.txt
# Important Requirements here
https://github.com/New-dev0/Telethon/archive/platy.zip
python-decouple
python-dotenv

# optional
akinator
apscheduler
aiohttp
bs4
enhancer>=0.3.4
git+https://github.com/New-dev0/instagrapi.git@39df1b1#egg=instagrapi
git+https://github.com/buddhhu/img2html.git@c44170d#egg=img2html
heroku3
gitpython
google-api-python-client
htmlwebshot
lottie
numpy>=1.21.2
oauth2client
opencv-python-headless
pillow>=7.0.0
profanitydetector
psutil
pypdf2>=1.26.0
pytz
qrcode
requests
tabulate
git+https://github.com/xditya/telegraph.git@41fd800#egg=telegraph
tgcrypto
youtube-search-python
yt-dlp

# extras
redis[hiredis]
pymongo[srv]
pyyaml>=5.4.1
pytgcalls>=3.0.0.dev21
aiofiles
coloredlogs

# addons
covid
emoji
fonttools
gtts
gingerit
jikanpy
git+https://github.com/New-dev0/AsyncLyricsExtractor.git
phlogo
pokedex.py
pyfiglet
pygments
pyjokes
quotefancy
shazamio
speechrecognition
speedtest-cli
textblob
wikipedia

# custom
pyrogram
pymediainfo
music_tag
pygithub
spotdl
imgur_python
uvloop
